requirements.txt
environ-config
faker
factory-boy
flask
flask_graphql
graphene
graphene-mongo
mongoengine
setuptools
